not final

+++ MiniProject2/TeacherService/src/main/java/dk/si/teachers/model/Teacher.java
@@ -14,7 +14,7 @@ import javax.persistence.*;
 public class Teacher {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @NonNull private Long sid;
+    @NonNull private int sid;
     @NonNull private String firstname;
     @NonNull private String lastname;
     @NonNull private Course course;
